+++ sprites/shot.py
@@ -4,11 +4,11 @@ from sprites.circleshape import CircleShape
 from constants import SHOT_RADIUS
 
 class Shot(CircleShape):
-    def __init__(self, x: int, y: int):
+    def __init__(self, x: float, y: float):
         super().__init__(x, y, SHOT_RADIUS)
 
     def draw(self, screen: pygame.display):
         pygame.draw.circle(screen, "white", self.position, self.radius, 2)
 
-    def update(self, dt: int):
+    def update(self, dt: float):
         self.position = self.position + (self.velocity * dt)
